requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
pytest==6.2.4
selenium~=4.1.3
pytest-assume==2.4.3
python-dotenv==0.19.2
pytest-xdist==2.5.0
Appium-Python-Client~=2.2.0
Faker~=13.11.0